Slightly more sane keymapping
Adjust keymapping to fit the Chip8's layout. We already got the 4x4 grid, just need to shift things around a bit. Normally, 0x2, 0x4, 0x6, and 0x8 are in a sort of dpad layout which is nice and convenient. For tic-tac-toe, the top left 3x3 section maps directly to the board. All of that is seriously messed up by the current layout, making the emulator very strange to use.